+++ cppcheck/tool.py
@@ -1,15 +1,17 @@
-from shutil import ExecError
 import subprocess
 import xmltodict
-import json
 import os
 import re
-from cestcrypto import dohash
 
 TOOL = "Cppcheckv 2.7"
 TOOL_FOLDER = "/pod-storage/{}/".format(TOOL.replace(" ", ""))
 INTERMEDIATE_RESULT = "/result/"
 READABLE_TOE = "/toe/"
+
+class FailedProcess:
+    def __init__(self, stderr):
+        self.stdout = ""
+        self.stderr = stderr
 
 
 def init_tool():
@@ -23,30 +25,30 @@ def run_tool(result_folder, argument, tools_are_silent):
     command = "/cppcheck/build/bin/cppcheck {} {}".format(args, READABLE_TOE)
     command_no_extra_spaces = re.sub(" +", " ", command)
 
-    # Save tool, version and command to file
-    tool_and_args = {
-        "tool": TOOL,
-        "args": [command_no_extra_spaces],
-    }
-
     # Run analysis
-    result = subprocess.run(
-        command_no_extra_spaces.split(" "), capture_output=True, text=True)
-
     try:
+        result = subprocess.run(
+            command_no_extra_spaces.split(" "), capture_output=True, text=True)
+    except Exception as e:
+        result = FailedProcess(str(e))
+
+    if os.path.exists(output_file):
         with open(output_file, "r") as output_file_bytes:
             output = output_file_bytes.read()
-    except:
-        raise Exception(result.stdout)
+        result_json = xmltodict.parse(output)
+    else:
+        result_json = {"error": "no output file found"}
 
-    result_json = xmltodict.parse(output)
+    
     full_report = {
         "tool": TOOL,
+        "port_version": 1,
         "command": [command_no_extra_spaces],
         "report": result_json,
     }
 
-    os.remove(output_file)
+    if os.path.exists(output_file):
+        os.remove(output_file)
 
     report_with_std = {
         "full_report": full_report,
