+++ dependency-check/tool.py
@@ -2,13 +2,16 @@ import subprocess
 import json
 import os
 import re
-from cestcrypto import dohash
 
 TOOL = "Dependency-Check 7.2.1"
 TOOL_FOLDER = "/pod-storage/{}/".format(TOOL.replace(" ", ""))
 INTERMEDIATE_RESULT = "/result/"
 READABLE_TOE = "/toe/"
 
+class FailedProcess:
+    def __init__(self, stderr):
+        self.stdout = ""
+        self.stderr = stderr
 
 def init_tool():
     return
@@ -23,17 +26,25 @@ def run_tool(result_folder, argument, tools_are_silent):
     command_no_extra_spaces = re.sub(" +", " ", command)
 
     # Run analysis
-    result = subprocess.run(command_no_extra_spaces.split(
-        " "), capture_output=tools_are_silent, text=True)
+    try: 
+        result = subprocess.run(command_no_extra_spaces.split(
+            " "), capture_output=tools_are_silent, text=True)
+    except Exception as e:
+        result = FailedProcess(str(e))
 
-    with open(INTERMEDIATE_RESULT + "dependency-check-report.json", "r") as f:
-        result_json = json.load(f)
 
-    # Tool specific clean up
-    os.remove(INTERMEDIATE_RESULT + "dependency-check-report.json")
+    if os.path.exists(INTERMEDIATE_RESULT + "dependency-check-report.json"):
+        with open(INTERMEDIATE_RESULT + "dependency-check-report.json", "r") as f:
+            result_json = json.load(f)
+
+        # Tool specific clean up
+        os.remove(INTERMEDIATE_RESULT + "dependency-check-report.json")
+    else:
+        result_json = {"error": "no output file found"}
 
     full_report = {
         "tool": TOOL,
+        "port_version": 1,
         "command": [command_no_extra_spaces],
         "report": result_json,
     }
